wip: start adding test
