+++ apps/svelte-kit/package.json
@@ -6,9 +6,10 @@
   "scripts": {
     "dev": "vite dev",
     "build": "vite build",
+    "preview": "vite preview",
     "prepare": "svelte-kit sync || echo ''",
     "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
-    "test": "pnpm /^test:.*/",
+    "test": "echo 'skip'",
     "test:e2e": "playwright test",
     "lint": "pnpm /^lint:.*/",
     "lint:eslint": "eslint .",
