+++ packages/auto-options/src/autoOptions.test.ts
@@ -1,0 +1,5 @@
+import { expect, test } from "vitest";
+
+test("autoOptions", () => {
+  expect(true).toBe(true);
+});
