+++ packages/auto-options/src/autoOptions.ts
@@ -2,34 +2,7 @@ import { parse, type AST } from "svelte/compiler";
 import { walk } from "zimmerframe";
 import MagicString from "magic-string";
 import type { VariableDeclarator } from "estree";
-
-// TODO: we have no types for the typescript AST yet, probably should pull in a library to get them
-interface Identifier extends AST.BaseNode {
-  type: "Identifier";
-  name: string;
-}
-
-interface TypeAnnotation {
-  type:
-    | "TSTypeReference"
-    // TODO: this is probably right, but I don't know yet
-    | "TSTypeLiteral";
-  typeName: Identifier;
-}
-
-interface InterfaceBody extends AST.BaseNode {
-  type: "TSInterfaceBody";
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any -- TODO: write a type for this
-  body: any[];
-}
-
-interface InterfaceDeclaration extends AST.BaseNode {
-  type: "TSInterfaceDeclaration";
-  id: Identifier;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any -- TODO: write a type for this
-  typeParmeters: any[];
-  body: InterfaceBody;
-}
+import type { InterfaceDeclaration, TypeAnnotation } from "./types";
 
 // In svelte web component land, even simple things such as exposing props as attributes have to be
 // manually configured using <svelte:options customElement={}/>
