+++ pnpm-workspace.yaml
@@ -45,4 +45,5 @@ catalog:
   typescript: 5.8.3
   typescript-eslint: 8.29.1
   vite: 6.2.6
+  vitest: 3.1.1
   zimmerframe: 1.1.2
